feat: add code quality tools and pipeline
Adds the following tools that are used for static code analysis (just a
term for tools that work on uncompiled source code.)

- Bandit
- Black
- Flake8

These tools can be installed using the `make install-dev` command in the
dev container.

This PR also adds some CI tools so these are enforced through PRs.

**Please Note** You are going to want to make sure that all of your changes in the
future are done through PR's and not by pushing to the `main` branch in
git or you risk pushing up changes that will block other's PRs.

Typically you would prevent any pushes to the main branch using the
**Branch Protection Rules** in Github.

Here I explain the tools I addded what they do and why you'd use them.

Bandit is used to identify easy to find security issues in your code.
I know you have codeQL also setup but it's fine and good practice to use
multiple scanners.

https://bandit.readthedocs.io/en/latest/

Black is a tool that will automatically format code so you have a
consistent formatting. It's opinonated but you want that, this ensures
that everyone who works in this produces code that looks the same.

https://black.readthedocs.io/en/stable/

This is a tool called a linter that checks source code for known bugs
like unused imports or variables that are unintentionally overwritten
and a whole bunch of other issues that could cause bugs in your code.

This won't get rid of all bugs but it will reduce the number of bugs
that can be found by automated tools.

There is also some configuration in the .flake8 file that ensures this
plays nice with Black

https://flake8.pycqa.org/en/latest/