feat: remove dupes, reduce IO, and shorten methods
Description

Use f strings to do string interpolation that allows for the removal of
the duplicate code that uses the pattern

```python

  if lang == "en":
    foo("bar_en")
  if lang == "fr":
    foo("bar_fr")
```

and allows you to replace it with:

```python
  foo(f"bar_{langl}")
```

Move custom_styles large string to a file to clean up app.py.

This just cleans up the code and the custom_styles variable only needs
to be set once in the app.

Pull beautiful soup parser to it's own method

Moves the large block of cleanup code to it's own method which makes the
html_convert method easier to read.

Removes some of the unneccessary IO by writting a string to file then
immediately reading it right back out to manipulate it, this should
speed up this conversion a bit.

Move the header and footer reading and parsing to the module level as it
only needs to happen once in the application and should speed up doc
conversion.